WIP: search and advanced search components.
Porting over search and advanced search boxes to components. Still work in progress for the rest of the page.
