+++ src/views/QueryView.vue
@@ -2,66 +2,10 @@
   <div>
     <div class="container mt-3">
       <h3>Search Data</h3>
-      <div class="input-group mb-3">
-        <select v-model="selectedColumn" class="form-select">
-          <option disabled value="">
-            Select Column
-          </option>
-          <option v-for="column in columns" :key="column" :value="column">
-            {{ column }}
-          </option>
-        </select>
-        <input
-          v-model="searchTerm"
-          type="text"
-          class="form-control"
-          placeholder="Search term"
-        >
-        <button class="btn btn-primary" @click="search">
-          Search
-        </button>
-        <button class="btn btn-secondary" @click="showAdvancedSearch = !showAdvancedSearch">
-          Advanced Search
-        </button>
-      </div>
-
+      <SearchBar @search="recievedSearchTerm" @show-advanced="recievedShowAdvancedSearch" />
       <div v-if="showAdvancedSearch" class="mt-3">
         <h4>Advanced Search</h4>
-        <div class="row">
-          <div class="col-md-4">
-            <input
-              v-model="advSearchValue1"
-              type="text"
-              class="form-control"
-              placeholder="Value 1"
-            >
-          </div>
-          <div class="col-md-4">
-            <select v-model="selectedCondition" class="form-select">
-              <option disabled value="">
-                Select Condition
-              </option>
-              <option v-for="condition in conditions" :key="condition" :value="condition">
-                {{ condition }}
-              </option>
-            </select>
-          </div>
-          <div class="col-md-4">
-            <input
-              v-model="advSearchValue2"
-              type="text"
-              class="form-control"
-              placeholder="Value 2"
-            >
-          </div>
-        </div>
-        <div class="row mt-2">
-          <div class="col-md-12">
-            <button class="btn btn-primary" @click="advancedSearch">
-              Advanced Search
-            </button>
-          </div>
-        </div>
+        <AdvancedSearch @advanced-terms="recievedAdvancedTerms" />
       </div>
 
       <div class="mt-3">
@@ -118,6 +62,8 @@
 
 <script setup>
   import { ref, computed } from 'vue';
+  import SearchBar from '../components/query/SearchBar.vue';
+  import AdvancedSearch from '../components/query/AdvancedSearch.vue';
   const isSelectedColumn = (column) => selectedColumns.value.includes(column);
 
   const toggleColumn = (column) => {
@@ -162,19 +108,20 @@
   const searchTerm = ref('');
   const searchResults = ref([]);
   const showAdvancedSearch = ref(false);
-  const selectedColumn = ref('');
-  const advSearchValue1 = ref('');
-  const advSearchValue2 = ref('');
-  const selectedCondition = ref('');
+  // const selectedColumn = ref('');
   const selectedColumns = ref([]);
 
-  const columns = [
-    'ledv', 'lesv', 'lsv', 'lvef', 'lvmass', 'redv', 'resv', 'rsv', 'rvef', 'scar',
-    // ... other columns
-    'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v'
-  ];
-
-  const conditions = ['<', '>', '=', '!='];
+  function recievedSearchTerm(s){
+    searchTerm.value = s;
+    console.log(s);
+    search();
+  }
+  function recievedShowAdvancedSearch(toggle){
+    showAdvancedSearch.value = toggle;
+  }
+  function recievedAdvancedTerms(terms){
+    console.log(terms)
+  }
 
   const search = () => {
     // Perform search logic here
